requirements.txt
anyio==3.6.2
autopep8==2.0.1
beanie==1.17.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.92.0
h11==0.14.0
httptools==0.5.0
idna==3.4
lazy-model==0.0.5
markdown-it-py==2.1.0
mdurl==0.1.2
motor==3.1.1
passlib==1.7.4
pycodestyle==2.10.0
pydantic==1.10.5
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==0.21.1
PyYAML==6.0
rich==13.3.1
sniffio==1.3.0
starlette==0.25.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
